move install tests
